+++ scripts/generateKeys.js
@@ -63,6 +63,13 @@ function makeKeys(rootPath) {
             keys[getModuleName(node)] = [...stateKeys, ...props];
           }
         }
+        if (isInitialStateDeclaration(node)) {
+          const literalNode = isOptionsDeclaration(node)
+            ? getLiteralNode(node)
+            : node;
+          const props = literalNode ? getPropsNames(literalNode, true) : [];
+          keys[getModuleName(node)] = props;
+        }
       });
     });
 
@@ -203,6 +210,17 @@ function isStateReturnDeclaration(node) {
   const escapedName = getEscapedName(node);
   return (
     kindName === "TypeAliasDeclaration" && /.+StateReturn$/.test(escapedName)
+  );
+}
+
+/**
+ * @param {import("ts-morph").Node<Node>} node
+ */
+function isInitialStateDeclaration(node) {
+  const kindName = node.getKindName();
+  const escapedName = getEscapedName(node);
+  return (
+    kindName === "TypeAliasDeclaration" && /.+InitialState$/.test(escapedName)
   );
 }
 
@@ -365,10 +383,7 @@ function reduceKeys(acc, [moduleName, array]) {
 
   const finalString = `${declaration} = ${value};\n`;
 
-  if (!moduleName.endsWith("State")) {
-    return `${acc}export ${finalString}`;
-  }
-  return `${acc}${finalString}`;
+  return `${acc}export ${finalString}`;
 }
 
 function log(...args) {
