--- src/CheckboxTw/CheckboxDescription.tsx
@@ -1,44 +1,0 @@
-import { cx } from "@renderlesskit/react";
-import { BoxHTMLProps, BoxOptions, useBox } from "reakit";
-import { createComponent, createHook } from "reakit-system";
-
-import { useTheme } from "../theme";
-import { CHECKBOX_DESCRIPTION_KEYS } from "./__keys";
-import { CheckboxStateReturn } from "./CheckboxState";
-
-export type CheckboxDescriptionOptions = BoxOptions &
-  Pick<CheckboxStateReturn, "size">;
-
-export type CheckboxDescriptionHTMLProps = BoxHTMLProps;
-
-export type CheckboxDescriptionProps = CheckboxDescriptionOptions &
-  CheckboxDescriptionHTMLProps;
-
-export const useCheckboxDescription = createHook<
-  CheckboxDescriptionOptions,
-  CheckboxDescriptionHTMLProps
->({
-  name: "CheckboxDescription",
-  compose: useBox,
-  keys: CHECKBOX_DESCRIPTION_KEYS,
-
-  useProps(options, htmlProps) {
-    const { size } = options;
-    const { className: htmlClassName, ...restHtmlProps } = htmlProps;
-
-    const theme = useTheme("checkboxNew");
-    const className = cx(
-      theme.description.base,
-      theme.description.size[size],
-      restHtmlProps,
-    );
-
-    return { className, ...restHtmlProps };
-  },
-});
-
-export const CheckboxDescription = createComponent({
-  as: "div",
-  memo: true,
-  useHook: useCheckboxDescription,
-});
