--- src/CheckboxTw/index.ts
@@ -1,2 +1,0 @@
-export * from "./CheckboxLabel";
-export * from "./CheckboxInput";
