+++ src/checkboxNew/__keys.ts
@@ -1,15 +1,18 @@
 // Automatically generated
-const CHECKBOX_STATE_KEYS = [
+export const CHECKBOX_STATE_KEYS = [
   "state",
-  "icon",
-  "label",
-  "description",
-  "value",
   "size",
   "isChecked",
   "isIndeterminate",
   "isUnchecked",
   "setState",
+] as const;
+export const USE_CHECKBOX_STATE_KEYS = [
+  "defaultState",
+  "state",
+  "onStateChange",
+  "size",
+  "value",
 ] as const;
 export const CHECKBOX_DESCRIPTION_KEYS = CHECKBOX_STATE_KEYS;
 export const CHECKBOX_ICON_KEYS = CHECKBOX_DESCRIPTION_KEYS;
