+++ src/checkboxNew/index.ts
@@ -1,6 +1,2 @@
-export * from "./Checkbox";
-export * from "./CheckboxDescription";
-export * from "./CheckboxIcon";
-export * from "./CheckboxInput";
 export * from "./CheckboxLabel";
-export * from "./CheckboxText";
+export * from "./CheckboxInput";
